Updated package to match refactor done in main application
