+++ app/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
 
     <queries>
         <intent>
-            <action android:name="com.datalogic.aladdinapp.data.model.EndlessService" />
+            <action android:name="com.datalogic.aladdin.data.model.EndlessService" />
         </intent>
     </queries>
