+++ app/src/main/java/com/datalogic/aladdin/interfaces/IScannerOutput.kt
@@ -1,4 +1,4 @@
-package com.datalogic.aladdinapp.interfaces
+package com.datalogic.aladdin.interfaces
 
 interface IScannerOutput {
     /*
